@@ -1,6 +1,6 @@
 {
   "name": "react-virtualized-pivot",
-  "version": "0.2.0-alpha",
+  "version": "0.2.1-alpha",
   "description": "React Virtualized Pivot",
   "main": "docs/umd/react-virtualized-pivot.js",
   "scripts": {
@@ -22,12 +22,13 @@
   "dependencies": {
     "node-sass": "^4.3.0",
     "papaparse": "^4.1.4",
-    "quick-pivot": "^2.2.0",
-    "react": "15.4.2",
-    "react-dom": "15.4.2",
+    "quick-pivot": "^2.2.3",
+    "react": "^16.0.0-beta.2",
+    "react-dom": "^16.0.0-beta.2",
+    "react-hot-loader": "^3.0.0-beta.7",
     "react-select-plus": "^1.0.0-rc.3.patch1",
     "react-sortablejs": "^1.3.1",
-    "react-virtualized": "^9.2.1",
+    "react-virtualized": "^9.9.0",
     "sass-loader": "^6.0.2",
     "sortablejs": "^1.5.1"
   },
