@@ -4,7 +4,7 @@ module.exports = [
 		exclude: /(node_modules|bower_components|public\/)/,
 		loader: "babel-loader",
 		query: {
-			presets:[ 'es2015', 'react', 'stage-2' ]
+			presets: ['es2015', 'stage-2', 'react'],
 		}
 	},
 	{
